added fixed delete bug

+++ models/product.js
@@ -51,6 +51,9 @@ module.exports = class Product {
     getProductsFromFile((products) => {
       const updatedProducts = products.filter(prod => prod.id!==id);
       const product = products.find(prod => prod.id ===id);
+      if (!product) {
+        return;
+      }
       fs.writeFile(p,JSON.stringify(updatedProducts),(err)=>{
         if (!err) {
           Cart.deleteProduct(id,product.price);
